Gestures dispatchers that defines specific events
The dispatcher listen to events from the gesture recogniser, or from the
gestures emulator if in debug mode.

It must be added as a script in the scene that uses its events, otherwise
it won't be registered to the gestures event source (which also must be
added to the scene).

IMPLEMENTATION NOTE: the events are made static so any script can register
an handler to them at any type during the Unity gameobject lifecycle (in
Awake, Start, Update, ...). This could introduce complexity when changing
the scene and registering new event handlers.
If this becomes the case, it would be better to migrate to a dependency
injection solution with dispatchers instantiated ad hoc and injected in
scripts that need to register handlers for the gesture events.