Flipping also Hydro's camera so it stays in the right place
The camera is not flipped directly, but requires a parent armor that
has a zero relative position from Hydro and that cancels the flip effect,
introducing an opposite flip.

Added this hack because flipping the child animator instead of Hydro
itself makes Hydro get stuck because of its collider that is not flipped.